+++ CHANGELOG.md
@@ -2,5 +2,21 @@
 
 ## [Unreleased]
 
-## [0.0.0-beta-1] - 2023-07-20
-- Added DayNightBanner widget to display a banner with the Sun and Moon based on the current time of day.
+## [0.0.0-beta-2] - 2023-07-20
+### Added
+- Initial release of the `daynightbanner` package.
+- Added the `DayNightBanner` widget that displays a day-night banner with a sun and moon icon based on the provided hour of the day.
+- Customizable properties for the `DayNightBanner` widget, including `sunImage`, `moonImage`, `sunSize`, `moonSize`, `backgroundImage`, `widthOfSunMoon`, `bannerHeight`, `backgroundImageHeight`, `decoration`, and `child`.
+- Support for automatically updating the banner based on the provided hour.
+
+### Changed
+- None.
+
+### Removed
+- None.
+
+### Fixed
+- None.
+
+### Security
+- None.